% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_descriptive_statistics.R
\name{create_descriptive_statistics}
\alias{create_descriptive_statistics}
\title{Descriptive Statistics}
\usage{
create_descriptive_statistics(input_data, output_path)
}
\arguments{
\item{input_data}{A tibble of input data}

\item{output_path}{The path of where output file has to be created}
}
\description{
Reads the input file provided by Anacap and creates a new file with descriptive statistics
}
\examples{
create_descriptive_statistics(input_data)
}
\keyword{stats}
